// src/Utils/types.ts
export interface FormData {
    fullName: string;
    email: string;
    profession: string;
    address: string;
    city: string;
    state: string;
    company: string;
    employer: string;
    role: string;
    location: string;
    start: string;
    finish: string;
    currentJob: string;
    skill1: string;
    skill2: string;
    skill3: string;
    skill4: string;
    skill5: string;
    schoolName: string;
    schoolLocation: string;
    degree: string;
    fieldOfStudy: string;
    graduationMonth: string;
    graduationStartYear:string;
    graduationCompleteYear:string;
    graduationYear: string;
    phoneNumber: string;
    linkedInUrl: string;
    githubUrl: string;
    twitterUrl: string;
    portfolioUrl: string;
    certification1: string;
    certification2: string;
    certification3: string;
  }
  